VLab Issue #25765 Spatial Display improvements (ellipse, selection,etc.)
Changed the process of drawing an ellipse in the Spatial Display to
involve clicking at the center point, dragging to a point that the edge
should intersect, and releasing (instead of clicking and releasing,
moving, and click and releasing again).

Fixed algorithm used to determine which hazard event should be selected
when the user clicks on the Spatial Display over multiple (overlapping)
events. It now selects the smallest (in area) of the events over which
the click occurred.

Added property to start-up config (with overriding within individual
settings configuration files) that indicates whether vertices or
bounding box points should take precedence when the user clicks and
drags at a point where both a modifiable polygon's vertex and a
draggable point from its bounding box lie. The default is to have the
bounding box point take precedence, as before, but for Prob_WFO the
vertex takes precedence.

Removed erroneously displayed error messages that indicated that events
could not be unlocked when they were saved or proposed and had not
already been locked. Such error messages made no sense because the only
reason the events could not be unlocked was because they hadn't been
locked in the first place, which was perfectly valid.

Change-Id: Ie42dfc5e561de302355eee0dec1163e187247a68