VLab Issue #45994 Configurable recommender dialog buttons
Added code to allow recommenders to specify that their parameter
gathering dialogs may use customized command buttons instead of the
default "Run" and "Cancel" buttons. The customizations include varying
the number and labels of the buttons; the specification of which button
is equivalent to "Cancel" (meaning the recommender does not run), or
specifying that no button allows this, so that the recommender run
cannot be cancelled; the specification of which button should be
considered to have been invoked when the user clicks the "X" (close)
button in the title bar; and the specification of which button is the
default button.

When the user dismisses a recommender dialog with customized buttons,
and s/he does not cancel it (if cancellation is even allowed by the
customized buttons, as discussed above), a new key-value pairing is
passed into the recommender's execute() method's dialog values
parameter, giving the identifier of the button that was chosen to
dismiss the dialog.

Modified the Ownership Tool to use customized buttons, in particular
avoiding cancellation when the dialog is coming up to ask the user if
s/he accepts the ownership change (since cancellation at that point
breaks ownership).

Removed the option for recommenders to specify whether or not to set the
user name and workstation of modified events; they always set these
fields now. (Previously they had this option because those fields were
being used to show the user who owned the hazard event, but ownership
has been moved to a generic attribute, as messing with those fields was
causing erroneous notifications of broken lock errors.)

Added "userName" and "workstation" attributes to the event set given to
recommenders, so that they know the current user and workstation. Also
added these attributes to the environment map passed to metadata
generators for the same reason.