VLab Issue #26712 Visibility of geometries on Spatial Display
Added the ability to specify within StartUpConfig a visual buffer to be
drawn around hazard events on the Spatial Display that do not have
visual features. The buffer's thickness and color may be provided, or if
none is given, no buffer is drawn. This allows such hazard events'
geometries to stand out better against similarly-colored backgrounds.

Added corresponding buffer thickness and color as optional properties of
visual features, to allow the latter to stand out better against similar
backgrounds as well.

Added a use-for-centering flag to visual features' properties so that a
single visible visual feature for a hazard event may be designated as
the one to be used for centering purposes when the hazard event is
selected and the Spatial Display is recentered to show the event.
Previously, only base geometries were used for such zooming. Base
geometries are still used as fallbacks for any events that do not have
visual features, and for those for which none of their visual features
have this flag set to True.

Fixed bug caused by the user repeatedly and rapidly moving a
manipulation point on an editable spatial entity while the movements
were being processed by the Swath Recommender.

Fixed bug causing a newly (re)issued Prob_* event to have an editable
duration, even when it should not.

Change-Id: I19e1a3792a5bd44744bdfa109dbff16dd8d85a27