VLab Issue #46566 Convective Recommender Updates
This update includes the following:

Use the last N<=5 ProbSevere centroids to calclate a motion vector for an automated hazard event
(Note:, this really blows up the uncertainty it seems. Also, it appears to uncheck the motion vector Automation if it updates while selected. Need to revire the latter)

If Automated event attribute is overridden and the forecaster sets it back to automated, the last values from ProbSevere should show up immediately.
(Note: not working yet for geometry)

Darrel found the scaling factor for object motion based on latitude, and it is applied