VLab Issue #46258 Fix revert to last saved for PHI events
Changed the way "Revert to last saved" is handled in terms of triggering
recommenders. The reversion is now marked as such when triggering a
recommender, so that the latter knows that reversion is what caused
whatever change it is handling.

Changed the Swath Recommender to handle reversion.

Cleaned up the interdependencies script for PHI events so that it
properly renders the various megawidget editable or read-only depending
upon automation level, activate and activateModify, etc.

Added the ability for recommenders to specify, if they so desire, that
their modifications to hazard events should not set said events'
modification flag to true. This is used by the Swath Recommender to
ensure that modifications it makes in response to a revert do not cause
bolding.

Fixed bug that could cause null pointer exceptions when checking for
hazard conflicts. This bug was introduced during the refactor to use
read-only hazard events in most places that was necessitated by the
locking code review.

Change-Id: I322135a5a9faa32f051ad3636d73a1b9ea962ea5