VLab Issue #44464 Fix annoying overwrite-high-res-geometry messages
Removed the display of annoying "Are you sure you want to overwrite the
high-resolution geometry?" messages that would appear at times when it
was already clear that the user wanted to do this, or when the user
should not really have a choice. This was popping up, for example, when
recommenders returned geometry changes for issued events.