VLab Issue #20595 Issues with visual features, and Swath Rec errors
Changed the process of a hazard event's type being set to include the
removing of any existing visual features for that event.

Added "type" as a first-class hazard attribute that can trigger a
recommender run (specified in the "modifyRecommenders" entry for the
appropriate event type within HazardTypes.py). This allows changes in
type but nothing else to trigger an event, which is required if visual
features are to be recreated upon type changes. For PHI, this means
triggering the Swath Recommender upon such changes.

Added code to make the Swath Recommender's creation of forecast
polygons, and thus everything created from them (e.g. the swath visual
feature) more robust by safely handling such polygons that end up being
invalid (due to the base geometry having one or more concave elements).

Fixed bug in visual features caused by serializing of their colors
losing some information due to the conversion of floats to bytes to save
space. This has been corrected by serializing their components as
floats, and not trying to save space; the space saved turned out to be
fairly minimal anyway. (This bug was causing the Swath Recommender to
run more than needed.)

Tweaked the buffer thicknesses for the Swath Recommender's visual
features, and changed the text labels of said features to be white.