VLab Issue #46736 Improvements to Swath Recommender
Added the ability to mark specific hazard attributes (on a
per-hazard-type basis) as "session" attributes, meaning they are not
persisted.

Similarly, added the ability to mark specific visual features as
non-persisting, so that they do not persist either.

Used these new capabilities to allow the Swath Recommender to react to
the toggling of the "Preview Grid" checkbox by toggling on or off the
appropriate visual features just as before, but do nothing else. The
event is not locked by these changes, nor is it bolded in the Console or
saved to the database. Each session (client) can have the preview grid
toggled on or off separately, without stepping on one another's toes.
(The only limitation is that if one user has the event locked, the
others will not be able to turn the grid on or off, just as before.)

Changed the Prob_* events' base and visual feature colors as per HWT
feedback.

Fixed a bug with Revert to Last Saved for Prob_* events that caused the
polygons to not be reverted correctly. Also fixed preview grid updating
behavior when reverting to last saved.

Change-Id: I9d62bd6d9c9e805f604aa6038a9ff312327c0b78