VLab Issue #44500
This code review contains updates to the International SIGMET as well as changes to MetaData files. Changes to MetaData files are to ensure no modifications to the hazard event is being made in MetaData generation. International SIGMET tasks are development of the XML formatter, improvements to the Volcanic Ash SIGMET workflow, and changes to formatted products. In addition, this code now implements number and series tracking via the database.

Change-Id: I94eb71c67d4b870dae3384ff181316f9eceaeea6