VLab Issue #48027 Miscellaneous Java bug fixes for HWT Year 3
Removed code that disallowed ENDED events to be changed to be ELAPSED
when merging in changes to events from recommender results.

Added code to ensure that an event that a recommender has said to save
to the history list is not saved to said list twice because its status
also changed. It will now be saved at most once to the history list as a
result of the recommender changing it.

Changed the Console to show nothing for the lock column for events in
history lists.

Fixed bug in the Spinner megawidget that sometimes caused it to be laid
out to be too small to show some of its possible values in its entry
field.