VLab Issue #48027 Genericized visual feature recommender triggers
Removed the specific "hazard event visual feature changed" recommender
trigger, and folded the previous functionality into the "hazard event
modification" trigger.  This allows a single recommender execution to be
triggered by changes to both the visual features and other attributes of
an event.

Removed use of hazard attribute to track which Prob_* visual features
are "base" features from which session-specific visual features are to
be derived. This was done because race conditions would occasionally
cause the visual features generated by the Convective Recommender to be
cast aside in favor of visual features from another source when the
recommender's event modifications were being merged into the session,
causing the visual features to be out of sync with the hazard attribute
and leading to exceptions being thrown.

Also fixed bug that very occasionally caused a null pointer exception
when the user clicked and dragged a modification point -- it's likely
that it could only occur if the drawable the user attempted to modify
disappeared between the mouse press and mouse drag.

Also added code to check to ensure that recommender execution results
do not contain invalid geometries for modified events.

Change-Id: I6be92ee88c7e7ef08d90fee9a49b46781e3e0052