VLab Issue #50032 Settings-specific filtering on hazard attributes
Added the ability to specify, on a per-settings basis, zero or more
filters that apply to any generic attributes of a hazard attribute that
hold strings as values. For example, a hazard event of a certain type
might have a generic attribute, "threshold", that holds either "at or
above" or "below" as its possible values. If users of a specific
settings wish, the settings configuration file can be altered to include
a filter for that attribute. If a hazard event does not have the
attribute in question, or its value is null, then the event is not
filtered out (this prevents events of other types, which might not have
the attribute in question, from being filtered merely for this reason;
they will generally be filtered out for other reasons, e.g. their type
is not wanted with the current settings).

This new capability is now used to filter out (hide) automated Prob_*
hazard events that the Convective Recommender has determined fall below
threshold. The Prob_WFO.py settings has been altered to include a filter
on "threshold". The Convective Recommender, in turn, now sets
"threshold" to "below" instead of using the PROPOSED hazard status in
order to hide below-threshold events.

Also got rid of unneeded notifyModelChanged() call in message handler,
and switched over to having Prob_WFO.py be the default (starting)
settings loaded for all perspectives.

Change-Id: I686fd84831fcc54165caa057474e74948d60d400